Prenotazione Service Implementation 2

+++ src/main/java/com/booklaunch/booklaunch/service/impl/PrenotazioneServiceImpl.java
@@ -6,6 +6,7 @@ import com.booklaunch.booklaunch.exception.enums.PrenotazioneEnum;
 import com.booklaunch.booklaunch.exception.enums.RichiestaEnum;
 import com.booklaunch.booklaunch.exception.enums.UtenteEnum;
 import com.booklaunch.booklaunch.exception.handler.ApiRequestException;
+import com.booklaunch.booklaunch.model.Prenotazione;
 import com.booklaunch.booklaunch.repository.PrenotazioneRepository;
 import com.booklaunch.booklaunch.repository.RichiestaRepository;
 import com.booklaunch.booklaunch.repository.UtenteRepository;
@@ -15,6 +16,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
+import java.time.temporal.ChronoUnit;
+import java.time.temporal.Temporal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -43,14 +46,54 @@ public class PrenotazioneServiceImpl implements PrenotazioneService {
 
     @Override
     public PrenotazioneDTO insertPrenotazione(PrenotazioneDTO prenotazioneDTO) {
-        return null;
+        LocalDate today = LocalDate.now();
+        if (prenotazioneRepository.existsByUtenteAndData(prenotazioneDTO.getId_utente(), prenotazioneDTO.getData_prenotazione()) != 0) {
+            Prenotazione prenotazione = new Prenotazione(prenotazioneDTO);
+            if ((ChronoUnit.DAYS.between(today, prenotazioneDTO.getData_prenotazione()) >= 1)) {
+                prenotazione.setColazione(prenotazioneDTO.getColazione());
+                prenotazione.setPranzo(prenotazioneDTO.getPranzo());
+                prenotazione.setCena(prenotazioneDTO.getCena());
+                prenotazione.setData_prenotazione(prenotazioneDTO.getData_prenotazione());
+
+                if (prenotazioneDTO.getCena() == null) {
+                    prenotazione.setSacchetto_pranzo(prenotazioneDTO.getSacchetto_pranzo());
+                } else {
+                    prenotazioneEnum = PrenotazioneEnum.getPrenotazioneEnumByMessageCode("PRE_SE");
+                    throw new ApiRequestException(prenotazioneEnum.getMessage());
+                }
+
+                if (prenotazioneDTO.getPranzo() == null) {
+                    prenotazione.setSacchetto_cena(prenotazioneDTO.getSacchetto_cena());
+                } else {
+                    prenotazioneEnum = PrenotazioneEnum.getPrenotazioneEnumByMessageCode("PRE_SE");
+                    throw new ApiRequestException(prenotazioneEnum.getMessage());
+                }
+
+                prenotazioneRepository.save(prenotazione);
+
+                return new PrenotazioneDTO(prenotazione);
+            } else {
+                prenotazioneEnum = PrenotazioneEnum.getPrenotazioneEnumByMessageCode("PRE_E");
+                throw new ApiRequestException(prenotazioneEnum.getMessage());
+            }
+        } else {
+            prenotazioneEnum = PrenotazioneEnum.getPrenotazioneEnumByMessageCode("PRE_AE");
+            throw new ApiRequestException(prenotazioneEnum.getMessage());
+        }
     }
 
     @Override
     public Boolean deletePrenotazione(Long id_prenotazione) {
         if (prenotazioneRepository.existsById(id_prenotazione)) {
-            prenotazioneRepository.deleteById(id_prenotazione);
-            return true;
+            Prenotazione prenotazione = prenotazioneRepository.findById(id_prenotazione).get();
+            if (!prenotazione.getCheck_richiesta()) {
+                prenotazioneRepository.deleteById(id_prenotazione);
+                return true;
+            } else {
+                prenotazioneEnum = PrenotazioneEnum.getPrenotazioneEnumByMessageCode("PRE_DM");
+                throw new ApiRequestException(prenotazioneEnum.getMessage());
+            }
+
         } else {
             prenotazioneEnum = PrenotazioneEnum.getPrenotazioneEnumByMessageCode("PRE_IDNE");
             throw new ApiRequestException(prenotazioneEnum.getMessage());
@@ -59,9 +102,9 @@ public class PrenotazioneServiceImpl implements PrenotazioneService {
 
     @Override
     public List<PrenotazioneDTO> findByData(LocalDate data_prenotazione) {
-        if(prenotazioneRepository.existsByData_prenotazione(data_prenotazione) != 0){
+        if (prenotazioneRepository.existsByData_prenotazione(data_prenotazione) != 0) {
             return prenotazioneRepository.findByData_prenotazione(data_prenotazione).stream().map(PrenotazioneDTO::new).collect(Collectors.toList());
-        }else{
+        } else {
             prenotazioneEnum = PrenotazioneEnum.getPrenotazioneEnumByMessageCode("PRE_NE");
             throw new ApiRequestException(prenotazioneEnum.getMessage());
         }
@@ -119,17 +162,56 @@ public class PrenotazioneServiceImpl implements PrenotazioneService {
 
     @Override
     public PrenotazioneDTO updatePrenotazione(PrenotazioneDTO prenotazioneDTO) {
-        return null;
+        LocalDate today = LocalDate.now();
+        if (prenotazioneRepository.existsById(prenotazioneDTO.getId())) {
+            Prenotazione prenotazione = prenotazioneRepository.findById(prenotazioneDTO.getId()).get();
+
+            if (!prenotazione.getCheck_richiesta()) {
+                if ((ChronoUnit.DAYS.between(today, prenotazione.getData_prenotazione()) >= 1)) {
+                    prenotazione.setColazione(prenotazioneDTO.getColazione());
+                    prenotazione.setPranzo(prenotazioneDTO.getPranzo());
+                    prenotazione.setCena(prenotazioneDTO.getCena());
+
+                    if (prenotazioneDTO.getCena() == null) {
+                        prenotazione.setSacchetto_pranzo(prenotazioneDTO.getSacchetto_pranzo());
+                    } else {
+                        prenotazioneEnum = PrenotazioneEnum.getPrenotazioneEnumByMessageCode("PRE_SE");
+                        throw new ApiRequestException(prenotazioneEnum.getMessage());
+                    }
+
+                    if (prenotazioneDTO.getPranzo() == null) {
+                        prenotazione.setSacchetto_cena(prenotazioneDTO.getSacchetto_cena());
+                    } else {
+                        prenotazioneEnum = PrenotazioneEnum.getPrenotazioneEnumByMessageCode("PRE_SE");
+                        throw new ApiRequestException(prenotazioneEnum.getMessage());
+                    }
+
+                    prenotazioneRepository.save(prenotazione);
+
+                    return new PrenotazioneDTO(prenotazione);
+                } else {
+                    prenotazioneEnum = PrenotazioneEnum.getPrenotazioneEnumByMessageCode("PRE_E");
+                    throw new ApiRequestException(prenotazioneEnum.getMessage());
+                }
+
+            } else {
+                prenotazioneEnum = PrenotazioneEnum.getPrenotazioneEnumByMessageCode("PRE_DM");
+                throw new ApiRequestException(prenotazioneEnum.getMessage());
+            }
+        }else {
+            prenotazioneEnum = PrenotazioneEnum.getPrenotazioneEnumByMessageCode("PRE_IDNE");
+            throw new ApiRequestException(prenotazioneEnum.getMessage());
+        }
     }
 
     @Override
     public List<PrenotazioneDTO> findByUtente(Long id_utente) {
-        if(!utenteRepository.existsById(id_utente)){
+        if (!utenteRepository.existsById(id_utente)) {
             utenteEnum = UtenteEnum.getUtenteEnumByMessageCode("UTE_IDNE");
             throw new ApiRequestException(utenteEnum.getMessage());
-        }else if(utenteRepository.existsById(id_utente)){
+        } else if (utenteRepository.existsById(id_utente)) {
             return prenotazioneRepository.prenotazioneByUtente(id_utente).stream().map(PrenotazioneDTO::new).collect(Collectors.toList());
-        }else{
+        } else {
             prenotazioneEnum = PrenotazioneEnum.getPrenotazioneEnumByMessageCode("PRE_NE");
             throw new ApiRequestException(prenotazioneEnum.getMessage());
         }
